// news-aggregator-fe/src/interfaces/i-axios.ts
export interface IDelete {
  url: string
  body?: any
  headers?: any
}

export interface IPost extends IDelete {
  body?: object
  query?: Record<string, any>
}

export interface IPostMultipart extends IDelete {
  data: FormData
}

export type IPatch = IPost

export type IPut = IPost

export interface IGet extends IDelete {
  query?: Record<string, any>
  body?: any
}

export interface IResponse<D> {
  data?: D
  code?: number
  message?: string
}

export interface RequestError {
  code: number,
  message: string
}